feat: refactor prediction setup and alerts flow
- Add PredictionSetup and ThresholdTilesExplorer components
- Add AlertPreviewPanel for prediction alerts
- Remove MarkReadyForForecastingModal and PredictionAlertsConfigurator
- Add periods and sortByCreated utilities
- Simplify PredictionDetailsGrid and dashboard widgets
