+++ apps/modeling-app/src/components/PageContent/ConfiguredModelDashboard/ConfiguredModelDashboard.tsx
@@ -1,6 +1,7 @@
 import { useMemo } from 'react';
 import type { PredictionInfo } from '@dhis2-chap/ui';
 import { useParams } from 'react-router-dom';
+import { sortByCreatedDesc } from '../../../utils/sortByCreated';
 import { JOB_STATUSES, useJobs } from '../../../hooks/useJobs';
 import { usePredictionSetup } from '../../../hooks/usePredictionSetup';
 import { ActivityWidget } from './ActivityWidget';
@@ -9,41 +10,27 @@ import { QuickActionsWidget } from './QuickActionsWidget';
 import { SummaryWidget } from './SummaryWidget';
 import styles from './ConfiguredModelDashboard.module.css';
 
-const getCreatedTime = (created?: string | null) => {
-    if (!created) {
-        return 0;
-    }
-
-    const time = Date.parse(created);
-    return Number.isNaN(time) ? 0 : time;
-};
-
-const getLatestPredictions = (predictions?: PredictionInfo[]) => (
-    [...(predictions || [])]
-        .sort((first, second) => getCreatedTime(second.created) - getCreatedTime(first.created))
-);
-
 export const ConfiguredModelDashboard: React.FC = () => {
-    const { configuredId } = useParams();
-    const parsedConfiguredId = Number(configuredId);
+    const { predictionSetupId } = useParams();
+    const parsedPredictionSetupId = Number(predictionSetupId);
     const {
         predictionSetup,
         error,
         hasValidPredictionSetupId,
         isLoading,
-    } = usePredictionSetup(configuredId);
+    } = usePredictionSetup(predictionSetupId);
     const {
         jobs = [],
         error: jobsError,
         isLoading: isLoadingJobs,
     } = useJobs({
-        predictionSetupId: parsedConfiguredId,
+        predictionSetupId: parsedPredictionSetupId,
         enabled: hasValidPredictionSetupId,
     });
 
     const predictions = useMemo(() => (
         hasValidPredictionSetupId
-            ? getLatestPredictions(predictionSetup?.predictions)
+            ? sortByCreatedDesc(predictionSetup?.predictions ?? [])
             : []
     ), [predictionSetup?.predictions, hasValidPredictionSetupId]);
     const hasRunningJob = useMemo(() => jobs.some(job => (
@@ -55,23 +42,23 @@ export const ConfiguredModelDashboard: React.FC = () => {
         <div className={styles.container}>
             <div className={styles.leftColumn}>
                 <PredictionRunsWidget
-                    configuredId={configuredId}
+                    predictionSetupId={predictionSetupId}
                     error={error}
-                    hasValidConfiguredId={hasValidPredictionSetupId}
+                    hasValidPredictionSetupId={hasValidPredictionSetupId}
                     hasRunningJob={hasRunningJob}
                     isLoading={isLoading}
                     predictions={predictions}
                 />
                 <ActivityWidget
                     error={jobsError}
-                    hasValidConfiguredId={hasValidPredictionSetupId}
+                    hasValidPredictionSetupId={hasValidPredictionSetupId}
                     isLoading={isLoadingJobs}
                     jobs={jobs}
                 />
             </div>
             <div className={styles.rightColumn}>
                 <QuickActionsWidget
-                    configuredId={configuredId}
+                    predictionSetupId={predictionSetupId}
                     predictionSetup={predictionSetup}
                     isLoading={isLoading}
                     latestPredictionId={predictions[0]?.id}
