+++ apps/modeling-app/src/components/PageContent/ConfiguredModelDashboard/QuickActionsWidget/QuickActionsWidget.tsx
@@ -7,67 +7,29 @@ import {
 import i18n from '@dhis2/d2-i18n';
 import { useState } from 'react';
 import { Widget } from '@dhis2-chap/ui';
-import type { PredictionSetupReadWithPredictions, QuantileTarget } from '@dhis2-chap/ui';
+import type { PredictionSetupReadWithPredictions } from '@dhis2-chap/ui';
 import { useNavigate } from 'react-router-dom';
-import { MarkReadyForForecastingModal } from '../../EvaluationDetails/QuickActionsWidget/MarkReadyForForecastingModal';
-import type { MarkReadyForForecastingFormValues } from '../../EvaluationDetails/QuickActionsWidget/MarkReadyForForecastingModal';
-import { getPredictionSetupQuantileTargets } from '@/utils/predictionSetupImportMapping';
+import {
+    MarkReadyForForecastingModal,
+    type MarkReadyForForecastingFormValues,
+} from '../../../PredictionSetup/MarkReadyForForecastingModal';
+import {
+    buildQuantileTargetsFromForm,
+    formValuesFromQuantileTargets,
+    getPredictionSetupQuantileTargets,
+} from '@/utils/predictionSetupImportMapping';
 import { useUpdatePredictionSetup } from './hooks/useUpdatePredictionSetup';
 import styles from './QuickActionsWidget.module.css';
 
 type Props = {
-    configuredId?: string;
+    predictionSetupId?: string;
     predictionSetup?: PredictionSetupReadWithPredictions;
     isLoading: boolean;
     latestPredictionId?: number;
 };
 
-const quantileKeys = [
-    'quantile_high',
-    'quantile_mid_high',
-    'median',
-    'quantile_mid_low',
-    'quantile_low',
-] as const;
-
-const getDataElementId = (
-    quantileTargets: QuantileTarget[],
-    quantileKey: typeof quantileKeys[number],
-) => (
-    quantileTargets.find(target => target.quantile === quantileKey)?.dataElementId ?? ''
-);
-
-const buildEditSetupFormValues = (
-    predictionSetup: PredictionSetupReadWithPredictions,
-): MarkReadyForForecastingFormValues => {
-    const quantileTargets = getPredictionSetupQuantileTargets(predictionSetup);
-
-    return {
-        name: predictionSetup.name,
-        use_import_mapping: quantileTargets.length > 0,
-        quantile_high: getDataElementId(quantileTargets, 'quantile_high'),
-        quantile_mid_high: getDataElementId(quantileTargets, 'quantile_mid_high'),
-        median: getDataElementId(quantileTargets, 'median'),
-        quantile_mid_low: getDataElementId(quantileTargets, 'quantile_mid_low'),
-        quantile_low: getDataElementId(quantileTargets, 'quantile_low'),
-    };
-};
-
-const buildQuantileTargets = (
-    values: MarkReadyForForecastingFormValues,
-): QuantileTarget[] => {
-    if (!values.use_import_mapping) {
-        return [];
-    }
-
-    return quantileKeys.map(quantile => ({
-        quantile,
-        dataElementId: values[quantile],
-    }));
-};
-
 export const QuickActionsWidget = ({
-    configuredId,
+    predictionSetupId,
     predictionSetup,
     isLoading,
     latestPredictionId,
@@ -75,25 +37,25 @@ export const QuickActionsWidget = ({
     const navigate = useNavigate();
     const [configurationModalIsOpen, setConfigurationModalIsOpen] = useState(false);
     const { updatePredictionSetup, isUpdating } = useUpdatePredictionSetup();
-    const canPredict = !!configuredId && !!predictionSetup?.configuredModel?.id;
+    const canPredict = !!predictionSetupId && !!predictionSetup?.configuredModel?.id;
 
     const handlePredict = () => {
         if (!predictionSetup) {
             return;
         }
 
-        const returnTo = `/predictions/${configuredId}`;
+        const returnTo = `/predictions/${predictionSetupId}`;
         navigate(
-            `/predictions/${configuredId}/new?returnTo=${encodeURIComponent(returnTo)}`,
+            `/predictions/${predictionSetupId}/new?returnTo=${encodeURIComponent(returnTo)}`,
         );
     };
 
     const handleShowLastRun = () => {
-        if (!configuredId || !latestPredictionId) {
+        if (!predictionSetupId || !latestPredictionId) {
             return;
         }
 
-        navigate(`/predictions/${configuredId}/runs/${latestPredictionId}`);
+        navigate(`/predictions/${predictionSetupId}/runs/${latestPredictionId}`);
     };
 
     const handleEditSetup = () => {
@@ -105,17 +67,22 @@ export const QuickActionsWidget = ({
             return;
         }
 
-        const quantileTargets = buildQuantileTargets(values);
-
         await updatePredictionSetup({
             predictionSetupId: predictionSetup.id,
             data: {
                 name: values.name,
-                quantileTargets,
+                quantileTargets: buildQuantileTargetsFromForm(values),
             },
         });
         setConfigurationModalIsOpen(false);
     };
+
+    const editSetupDefaultValues = predictionSetup
+        ? formValuesFromQuantileTargets(
+                predictionSetup.name,
+                getPredictionSetupQuantileTargets(predictionSetup),
+            )
+        : undefined;
 
     return (
         <>
@@ -162,7 +129,7 @@ export const QuickActionsWidget = ({
                 <MarkReadyForForecastingModal
                     onClose={() => setConfigurationModalIsOpen(false)}
                     onSubmit={handleEditSetupSubmit}
-                    defaultValues={buildEditSetupFormValues(predictionSetup)}
+                    defaultValues={editSetupDefaultValues}
                     title={i18n.t('Edit prediction setup')}
                     isSubmitting={isUpdating}
                 />
