+++ apps/modeling-app/src/components/PageContent/EvaluationDetails/QuickActionsWidget/QuickActionsWidget.tsx
@@ -3,10 +3,12 @@ import { Button, IconCheckmarkCircle16, IconVisualizationLineMulti16, IconDuplic
 import i18n from '@dhis2/d2-i18n';
 import { useNavigate } from 'react-router-dom';
 import { Widget } from '@dhis2-chap/ui';
-import type { QuantileTarget } from '@dhis2-chap/ui';
 import { CopyBacktestModal } from '../../../BacktestsTable/BacktestActionsMenu/CopyBacktestModal';
-import { MarkReadyForForecastingModal } from './MarkReadyForForecastingModal';
-import type { MarkReadyForForecastingFormValues } from './MarkReadyForForecastingModal';
+import {
+    MarkReadyForForecastingModal,
+    type MarkReadyForForecastingFormValues,
+} from '../../../PredictionSetup/MarkReadyForForecastingModal';
+import { buildQuantileTargetsFromForm } from '@/utils/predictionSetupImportMapping';
 import { useCreatePredictionSetup } from './hooks/useCreatePredictionSetup';
 import styles from './QuickActionsWidget.module.css';
 
@@ -52,23 +54,11 @@ export const QuickActionsWidget = ({
     };
 
     const handleMarkReadySubmit = async (values: MarkReadyForForecastingFormValues) => {
-        const quantileTargets: QuantileTarget[] = [];
-
-        if (values.use_import_mapping) {
-            quantileTargets.push(
-                { quantile: 'quantile_high', dataElementId: values.quantile_high },
-                { quantile: 'quantile_mid_high', dataElementId: values.quantile_mid_high },
-                { quantile: 'median', dataElementId: values.median },
-                { quantile: 'quantile_mid_low', dataElementId: values.quantile_mid_low },
-                { quantile: 'quantile_low', dataElementId: values.quantile_low },
-            );
-        }
-
         await createPredictionSetup({
             data: {
                 backtestId: evaluationId,
                 name: values.name,
-                quantileTargets,
+                quantileTargets: buildQuantileTargetsFromForm(values),
             },
         });
         setMarkReadyModalIsOpen(false);
