+++ apps/modeling-app/src/components/ThresholdTilesExplorer/index.ts
@@ -1,0 +1,6 @@
+export {
+    OutbreakProbabilityControl,
+    SummaryRow,
+    ThresholdTilesExplorer,
+    type StatusFilterValue,
+} from './ThresholdTilesExplorer';
